Moved Api folder into services folder
