+++ src/main/java/services/api/dtos/DroneType.java
@@ -1,4 +1,4 @@
-package main.java.api.dtos;
+package main.java.services.api.dtos;
 
 public class DroneType {
